# fix a bug, made some test and refactor a bit