I think this time I got it.
Made an integrity test - I generated random file based on a seed and ssdeeped them with the official implementation. The test generating the files and ssdeep them and then comparing results.
I refactored a bit, so the bytes and the file methods now use the same core code.