@glaslos 2017-12-11
Can you remove this line: https://github.com/glaslos/ssdeep/blob/master/.travis.yml#L4 or fix the backwards compatibility issue?