@glaslos 2017-12-11
Could you also add a `make benchcmp` as a comment?