@davidt99 2017-12-12
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkDistance-8        9101          9419          +3.49%
BenchmarkRollingHash-8     3.03          3.19          +5.28%
BenchmarkSumHash-8         33.4          34.2          +2.40%
BenchmarkBlockSize-8       11.1          6.70          -39.64%
BenchmarkProcessByte-8     29.7          28.6          -3.70%
```
Not sure why there is a performance improvement for GetBlockSize - It didn't change.