@glaslos 2017-12-12
Yeah, no clue why there is that performance improvement. I was running a benchcmp as well and could reproduce your observation. Thanks for fixing this issue! I will create a new release later today.